pyproject.toml
[project]
name = "hermes-overseer"
version = "0.1.0"
description = "Off-device overseer for hermes-agent VPS"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
overseer = "overseer.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/overseer"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "respx>=0.22",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["SIM108", "UP046"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
